requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==2.3.7
flask-mongoengine==1.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
pymongo==4.6.0
